refactor: improve code structure and readability
