+++ src/components/Forms/Admin/Form.module.scss
@@ -17,11 +17,11 @@
   box-shadow: 2px 3px 37px 0 rgb(0 0 0 / 25%);
 
   &__add-book {
-    margin-top: 165px;
+    margin-top: 365px;
   }
 
   &__edit-book {
-    margin-top: 268px;
+    margin-top: 465px;
   }
 
   .title-container {
@@ -90,6 +90,9 @@ form {
   .input-container {
     display: flex;
     gap: 14px;
+    align-items: end;
+    justify-content: space-around;
+
     width: 100%;
   }
