+++ src/components/Forms/Admin/types.ts
@@ -1,0 +1,8 @@
+import { BookById } from '@/redux/services/services.types';
+
+export interface EditBookProps {
+  id: number;
+  book: BookById;
+  handleClose: () => void;
+  handleOpenPopup: () => void;
+}
